examples/mcp_servers/telemetry_passback/pyproject.toml
[project]
name = "telemetry-passback"
version = "0.1.0"
description = "SEP-0000 reference implementation: cross-org distributed tracing via MCP serverExecutionTelemetry"
requires-python = ">=3.11"
dependencies = [
    "arcade-mcp-server>=1.17.4,<2.0.0",
    "mcp>=1.19.0",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langgraph",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-exporter-otlp-proto-http",
    "opentelemetry-instrumentation-langchain",
    "opentelemetry-instrumentation-httpx",
    "httpx",
    "python-dotenv",
    "protobuf",
]

[project.optional-dependencies]
dev = [
    "arcade-mcp[all]>=1.11.2,<2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telemetry_passback"]

# Use local arcade-mcp-server (TelemetryPassbackMiddleware is not yet published)
[tool.uv.sources]
arcade-mcp-server = { path = "../../../libs/arcade-mcp-server/", editable = true }
arcade-serve = { path = "../../../libs/arcade-serve/", editable = true }
